Make small config constants int

+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/config/SpecConfigBuilder.java
@@ -988,8 +988,8 @@ public class SpecConfigBuilder {
     private Bytes4 eip4844ForkVersion;
     private UInt64 eip4844ForkEpoch;
 
-    private UInt64 fieldElementsPerBlob;
-    private UInt64 maxBlobsPerBlock;
+    private int fieldElementsPerBlob;
+    private int maxBlobsPerBlock;
 
     private Eip4844Builder() {}
 
@@ -1016,12 +1016,12 @@ public class SpecConfigBuilder {
       return this;
     }
 
-    public Eip4844Builder fieldElementsPerBlob(final UInt64 fieldElementsPerBlob) {
+    public Eip4844Builder fieldElementsPerBlob(final int fieldElementsPerBlob) {
       this.fieldElementsPerBlob = fieldElementsPerBlob;
       return this;
     }
 
-    public Eip4844Builder maxBlobsPerBlock(final UInt64 maxBlobsPerBlock) {
+    public Eip4844Builder maxBlobsPerBlock(final int maxBlobsPerBlock) {
       this.maxBlobsPerBlock = maxBlobsPerBlock;
       return this;
     }
