+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/datastructures/blocks/blockbody/versions/eip4844/BlindedBeaconBlockBodySchemaEip4844Impl.java
@@ -135,7 +135,7 @@ public class BlindedBeaconBlockBodySchemaEip4844Impl
         namedSchema(
             BlockBodyFields.BLOB_KZG_COMMITMENTS,
             SszListSchema.create(
-                SszKZGCommitmentSchema.INSTANCE, specConfig.getMaxBlobsPerBlock().longValue())));
+                SszKZGCommitmentSchema.INSTANCE, specConfig.getMaxBlobsPerBlock())));
   }
 
   @Override
